fixed the tie breaker bug
my highest_card was in the wrong place and was constantly reset to 0 every player iteration. moving it out and before the player loop fixed it. also noticed a problem in the elim_player function: captured deck was not converting to battle deck in the event that the player's hand was not empty, so even in a strategic battle the players eventually only had 1 card in their hand during a round and as a result went to "top-decking" only.